Refactor CLI runtime wiring out of library entry
