+++ context.md
@@ -18,7 +18,7 @@
 ## Key Entry Points
 
 - `index.js`: re-exports the library API from `src/lib/index.js` and forwards direct execution to the CLI runner.
-- `src/lib/index.js`: bootstraps the agent loop, exposes helpers for tests, and provides startup flag utilities (`setStartupFlags`, `applyStartupFlagsFromArgv`).
+- `src/lib/index.js`: aggregates CLI-agnostic helpers, re-exporting the CLI runtime without owning its rendering logic.
 - `src/cli/runner.js`: handles `templates`/`shortcuts` subcommands and wires CLI startup flags before launching the loop.
 - `src/agent/loop.js`: orchestrates OpenAI interactions, now delegating each pass to `passExecutor.js` while managing readline/ESC state.
 - `src/commands/run.js`: sandbox for process execution with timeout + cancellation; re-exports specialised helpers (browse/read/edit/replace/escape-string).
@@ -61,7 +61,7 @@
 
 ## Key Entry Points
 
-- `index.js`: bootstraps the agent loop, handles `templates`/`shortcuts` subcommands, exposes helpers for tests, honours startup flags (`--auto-approve`, `--nohuman`).
+- `index.js`: delegates CLI launches to `src/cli/runner.js`, keeps the library surface re-exported for programmatic consumers.
 - `src/agent/loop.js`: orchestrates OpenAI interactions, now delegating each pass to `passExecutor.js` while managing readline/ESC state.
 - `src/commands/run.js`: sandbox for process execution with timeout + cancellation; re-exports specialised helpers (browse/read/edit/replace/escape-string).
