+++ src/lib/context.md
@@ -6,9 +6,10 @@
 
 ## Modules
 
-- `index.js`: re-exports command runners, prompt helpers, and startup flag utilities while wiring the agent loop used by the CLI.
+- `index.js`: re-exports command runners, prompt helpers, and startup flag utilities without owning CLI rendering concerns.
+- `startupFlags.js`: centralises mutable startup flag state so the CLI runtime and library exports stay in sync.
 
 ## Notes
 
 - Startup flags default to `false` and can be configured via `setStartupFlags`, `parseStartupFlagsFromArgv`, or `applyStartupFlagsFromArgv`.
-- The CLI runner (`../cli/runner.js`) is the only module responsible for parsing process arguments and invoking `agentLoop()`.
+- The CLI runner (`../cli/runner.js`) is the only module responsible for parsing process arguments and invoking the CLI runtime.
